requirements.txt
streamlit
easyocr
Pillow
numpy